fix: corrected an issue where editor component height would be greater than current screen height
The editor component would cause a scrollbar due to not having a proper height limit based on
current screen height. I found out that it had 100dvh which caused a leak/scrollbar. So after
deleting that and adding a display: flex + flex: 1 settings it now fits perftly inside the current
screen height. The only scrollbar at the moment is inside the textarea which is currently
acceptable.