feat: added the ability to save all changes the current document using the save changes button
There was a problem with how the document saved individual changes so I reworked the design to make
it where currentDocument is the main source of data and all changes will be applied to that. This in
fact will not update myDocuments or be saved. In order to save the currentDocument, the user must
press the save changes button. Once this is done all changes will now update myDocuments with the
updated document. If the user does not save changes and opens a new document, all changes will be
lost.